API and tests for node installations
- Cleaned up the API for room/node installations.
- Added tests.
- Added documentation.
- Minor cleanup.
